Implement the spinner selection from the main app
Device selection should still be deleted and replaced totally by the
spinner selection from the main activity.
The first data load from the device should still be handled.
For now a saved deviceID is used to retrieve the last used device. In
case the app was never used, this ID is not available.