Refactor and design improvement.
All device are now loaded after login.
No more device selection screen, we directly jump to the main activity.
If the previously used device is still available, we'll control that one. If he is not, the first device of the list will be used as active device.
A Parcelable object representing the device and his actual color set is created. Allowing to pass the device and their respective info from one activity to the other.